+++ modules/frontroute/src/main/scala/io/frontroute/PathMatcher.scala
@@ -162,7 +162,9 @@ trait PathMatchers {
 
   def double: PathMatcher[Double] = segment.tryParse(_.toDouble)
 
-  implicit def stringToSegment(s: String): PathMatcher[Unit]    = segment(s)
-  implicit def regexToPathMatcher(r: Regex): PathMatcher[Match] = regex(r)
+  implicit def stringToSegment(s: String): PathMatcher[Unit]         = segment(s)
+  implicit def setToSegment(oneOf: Set[String]): PathMatcher[String] = segment(oneOf)
+  implicit def setToSegment(oneOf: Seq[String]): PathMatcher[String] = segment(oneOf)
+  implicit def regexToPathMatcher(r: Regex): PathMatcher[Match]      = regex(r)
 
 }