+++ project/Dependencies.scala
@@ -30,6 +30,7 @@ object Dependencies {
       "io.laminext" %%% "markdown"               % DependencyVersions.laminext,
       "io.laminext" %%% "ui"                     % DependencyVersions.laminext,
       "io.laminext" %%% "tailwind"               % DependencyVersions.laminext,
+      "io.laminext" %%% "util"                   % DependencyVersions.laminext,
       "io.laminext" %%% "tailwind-default-theme" % DependencyVersions.laminext
     )
   }