+++ website/src/main/scala/io/frontroute/site/TemplateVars.scala
@@ -3,7 +3,7 @@ package io.frontroute.site
 object TemplateVars {
 
   val vars = Seq(
-    "frontrouteVersion" -> "0.13.3"
+    "frontrouteVersion" -> "0.14.0"
   )
 
   def apply(s: String): String =